Remove base class got HostEnvironment based layout renders
