+++ src/Shared/LayoutRenderers/AspNetAppBasePathLayoutRenderer.cs
@@ -7,6 +7,7 @@ using IHostEnvironment = Microsoft.AspNetCore.Hosting.IHostingEnvironment;
 #if ASP_NET_CORE3
 using Microsoft.Extensions.Hosting;
 #endif
+using NLog.Web.DependencyInjection;
 #else
 using NLog.Web.Internal;
 #endif
@@ -27,8 +28,38 @@ namespace NLog.Web.LayoutRenderers
 #endif
     [LayoutRenderer("aspnet-appbasepath")]
     [ThreadAgnostic]
-    public class AspNetAppBasePathLayoutRenderer : AspNetHostEnvironmentLayoutRendererBase
+    public class AspNetAppBasePathLayoutRenderer : LayoutRenderer
     {
+#if ASP_NET_CORE
+        /// <summary>
+        /// Context for DI
+        /// </summary>
+        private IHostEnvironment _hostEnvironment;
+
+        /// <summary>
+        /// Provides access to the current IHostEnvironment
+        /// </summary>
+        /// <returns>IHostEnvironment or <c>null</c></returns>
+        [NLogConfigurationIgnoreProperty]
+        public IHostEnvironment HostEnvironment
+        {
+            get => _hostEnvironment ?? (_hostEnvironment = RetrieveHostEnvironment(ResolveService<IServiceProvider>(), LoggingConfiguration));
+            set => _hostEnvironment = value;
+        }
+
+        internal static IHostEnvironment RetrieveHostEnvironment(IServiceProvider serviceProvider, LoggingConfiguration loggingConfiguration)
+        {
+            return ServiceLocator.ResolveService<IHostEnvironment>(serviceProvider, loggingConfiguration);
+        }
+#else
+        /// <summary>
+        /// Provides access to the current IHostEnvironment
+        /// </summary>
+        /// <returns>IHostEnvironment or <c>null</c></returns>
+        [NLogConfigurationIgnoreProperty]
+        public IHostEnvironment HostEnvironment { get; set; } = new HostEnvironment();
+#endif
+
         /// <inheritdoc />
         protected override void Append(StringBuilder builder, LogEventInfo logEvent)
         {
