+++ src/Shared/LayoutRenderers/IISInstanceNameLayoutRenderer.cs
@@ -8,10 +8,12 @@ using NLog.Web.Internal;
 #else
 #if ASP_NET_CORE2
 using Microsoft.AspNetCore.Hosting;
+using IHostEnvironment = Microsoft.AspNetCore.Hosting.IHostingEnvironment;
 #endif
 #if ASP_NET_CORE3
 using Microsoft.Extensions.Hosting;
 #endif
+using NLog.Web.DependencyInjection;
 #endif
 
 namespace NLog.Web.LayoutRenderers
@@ -28,8 +30,37 @@ namespace NLog.Web.LayoutRenderers
     [LayoutRenderer("iis-site-name")]
     // ReSharper disable once InconsistentNaming
     [ThreadAgnostic]
-    public class IISInstanceNameLayoutRenderer : AspNetHostEnvironmentLayoutRendererBase
+    public class IISInstanceNameLayoutRenderer : LayoutRenderer
     {
+#if ASP_NET_CORE
+        /// <summary>
+        /// Context for DI
+        /// </summary>
+        private IHostEnvironment _hostEnvironment;
+
+        /// <summary>
+        /// Provides access to the current IHostEnvironment
+        /// </summary>
+        /// <returns>IHostEnvironment or <c>null</c></returns>
+        [NLogConfigurationIgnoreProperty]
+        public IHostEnvironment HostEnvironment
+        {
+            get => _hostEnvironment ?? (_hostEnvironment = RetrieveHostEnvironment(ResolveService<IServiceProvider>(), LoggingConfiguration));
+            set => _hostEnvironment = value;
+        }
+
+        internal static IHostEnvironment RetrieveHostEnvironment(IServiceProvider serviceProvider, LoggingConfiguration loggingConfiguration)
+        {
+            return ServiceLocator.ResolveService<IHostEnvironment>(serviceProvider, loggingConfiguration);
+        }
+#else
+        /// <summary>
+        /// Provides access to the current IHostEnvironment
+        /// </summary>
+        /// <returns>IHostEnvironment or <c>null</c></returns>
+        [NLogConfigurationIgnoreProperty]
+        public IHostEnvironment HostEnvironment { get; set; } = new HostEnvironment();
+#endif
         /// <inheritdoc />
         protected override void Append(StringBuilder builder, LogEventInfo logEvent)
         {
